replace derivative with educe
