+++ Cargo.toml
@@ -28,6 +28,7 @@ mio = { version = "1.0", default-features = false }
 
 cfg-if = "1.0.0"
 polling = "2.8.0"
+educe = "0.6.0"
 
 libc = "0.2"
 rand = "0.8"
@@ -50,7 +51,6 @@ once_cell = "1"
 dashmap = "6"
 num_cpus = "1"
 uuid = "1"
-derivative = "2"
 tempfile = "3"
 cc = "1"
 syn = "2"
