+++ core/Cargo.toml
@@ -29,7 +29,7 @@ uuid = { workspace = true, features = [
     "v4",
     "fast-rng",
 ], optional = true }
-derivative = { workspace = true, optional = true }
+educe = { workspace = true, optional = true }
 core_affinity = { workspace = true, optional = true }
 crossbeam-utils = { workspace = true, optional = true }
 psm.workspace = true
@@ -77,7 +77,7 @@ backtrace.workspace = true
 log = ["tracing", "tracing-subscriber", "time"]
 
 # low-level raw coroutine
-korosensei = ["corosensei", "uuid", "nix/pthread", "derivative"]
+korosensei = ["corosensei", "uuid", "nix/pthread", "educe"]
 
 # Provide preemptive scheduling implementation.
 # Enable for default.
