+++ core/src/net/event_loop.rs
@@ -19,8 +19,9 @@ use std::time::Duration;
 
 cfg_if::cfg_if! {
     if #[cfg(all(target_os = "linux", feature = "io_uring"))] {
-        use libc::{epoll_event, iovec, msghdr, off_t, size_t, sockaddr, socklen_t, ssize_t};
+        use libc::{epoll_event, iovec, msghdr, off_t, size_t, sockaddr, socklen_t};
         use dashmap::DashMap;
+        use std::ffi::c_longlong;
     }
 }
 
@@ -36,7 +37,7 @@ pub(crate) struct EventLoop<'e> {
     operator: crate::net::operator::Operator<'e>,
     #[allow(clippy::type_complexity)]
     #[cfg(all(target_os = "linux", feature = "io_uring"))]
-    syscall_wait_table: DashMap<usize, Arc<(Mutex<Option<ssize_t>>, Condvar)>>,
+    syscall_wait_table: DashMap<usize, Arc<(Mutex<Option<c_longlong>>, Condvar)>>,
     selector: Poller,
     pool: CoroutinePool<'e>,
     phantom_data: PhantomData<&'e EventLoop<'e>>,
@@ -223,30 +224,9 @@ impl<'e> EventLoop<'e> {
             }
         }
 
-        #[cfg(all(target_os = "linux", feature = "io_uring"))]
-        if crate::net::operator::support_io_uring() {
-            // use io_uring
-            let (count, mut cq, left) = self.operator.select(left_time, 0)?;
-            if count > 0 {
-                for cqe in &mut cq {
-                    let token = usize::try_from(cqe.user_data()).expect("token overflow");
-                    if crate::common::constants::IO_URING_TIMEOUT_USERDATA == token {
-                        continue;
-                    }
-                    // resolve completed read/write tasks
-                    let result = cqe.result() as ssize_t;
-                    eprintln!("io_uring finish {token} {result}");
-                    if let Some((_, pair)) = self.syscall_wait_table.remove(&token) {
-                        let (lock, cvar) = &*pair;
-                        let mut pending = lock.lock().expect("lock failed");
-                        *pending = Some(result);
-                        cvar.notify_one();
-                    }
-                    unsafe { self.resume(token) };
-                }
-            }
-            if left != left_time {
-                left_time = Some(left.unwrap_or(Duration::ZERO));
+        cfg_if::cfg_if! {
+            if #[cfg(all(target_os = "linux", feature = "io_uring"))] {
+                left_time = self.adapt_io_uring(left_time)?;
             }
         }
 
@@ -261,6 +241,36 @@ impl<'e> EventLoop<'e> {
             }
         }
         Ok(())
+    }
+
+    #[cfg(all(target_os = "linux", feature = "io_uring"))]
+    fn adapt_io_uring(&self, mut left_time: Option<Duration>) -> std::io::Result<Option<Duration>> {
+        if crate::net::operator::support_io_uring() {
+            // use io_uring
+            let (count, mut cq, left) = self.operator.select(left_time, 0)?;
+            if count > 0 {
+                for cqe in &mut cq {
+                    let token = usize::try_from(cqe.user_data()).expect("token overflow");
+                    if crate::common::constants::IO_URING_TIMEOUT_USERDATA == token {
+                        continue;
+                    }
+                    // resolve completed read/write tasks
+                    let result = cqe.result() as c_longlong;
+                    eprintln!("io_uring finish {token} {result}");
+                    if let Some((_, pair)) = self.syscall_wait_table.remove(&token) {
+                        let (lock, cvar) = &*pair;
+                        let mut pending = lock.lock().expect("lock failed");
+                        *pending = Some(result);
+                        cvar.notify_one();
+                    }
+                    unsafe { self.resume(token) };
+                }
+            }
+            if left != left_time {
+                left_time = Some(left.unwrap_or(Duration::ZERO));
+            }
+        }
+        Ok(left_time)
     }
 
     #[allow(clippy::unused_self)]
@@ -404,7 +414,7 @@ macro_rules! impl_io_uring {
             pub(super) fn $syscall(
                 &self,
                 $($arg: $arg_type),*
-            ) -> std::io::Result<Arc<(Mutex<Option<ssize_t>>, Condvar)>> {
+            ) -> std::io::Result<Arc<(Mutex<Option<c_longlong>>, Condvar)>> {
                 let token = EventLoop::token(Syscall::$syscall);
                 self.operator.$syscall(token, $($arg, )*)?;
                 let arc = Arc::new((Mutex::new(None), Condvar::new()));
