+++ hook/src/lib.rs
@@ -68,11 +68,7 @@ pub(crate) fn hook() -> bool {
     clippy::similar_names,
     clippy::not_unsafe_ptr_arg_deref,
     clippy::many_single_char_names,
-    clippy::cast_sign_loss,
-    clippy::cast_possible_truncation,
-    clippy::cast_possible_wrap,
-    clippy::unnecessary_cast,
-    trivial_numeric_casts
+    clippy::unnecessary_cast
 )]
 pub mod syscall;
